home_node/backend/fastapi/requirements.txt
fastapi
uvicorn
#pydantic
bcrypt
redis
#databases
databases[sqlite]
python-jose[cryptography]